pyproject.toml
[tool.poetry]
name = "рabit-control-api"
version = "0.1.0"
description = ""
authors = ["Vladislav_Poroshin <vld.poroshin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
psycopg2-binary = "^2.9.9"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.0"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
redis = "^5.0.1"
djangorestframework-simplejwt = "^5.3.0"
celery = "^5.3.4"
django-celery-beat = "^2.5.0"


[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
